requirements.txt
fastapi
uvicorn
mangum
pydantic
qrcode
pillow
requests
whois
yt-dlp
user-agents
zxcvbn
python-multipart
